% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.osm.lines.R
\name{add.lines}
\alias{add.lines}
\title{add.osm.lines}
\usage{
add.lines(lines = lines, col = "gray40")
}
\arguments{
\item{lines}{= a list of lines returned by get.osm.polygons}

\item{col}{= colour of lines (default = "gray40")}
}
\value{
nothing (adds to graphics.device opened with plot.osm.polygons)
}
\description{
plot.osm.polygons initiates a graphics device to which add.osm.lines () and
add.osm.polygons () adds successive layers of lines and polygons generated by
get.osm.polygons ().
}

