% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.osm.polygons.R
\name{add.osm.polygons}
\alias{add.osm.polygons}
\title{add.osm.polygons}
\usage{
add.osm.polygons(poly = poly, col = "gray40")
}
\arguments{
\item{poly}{= a list of polygons returned by get.osm.polygons}

\item{col}{= colour of polygons (default = "gray40")}
}
\value{
nothing (adds to graphics.device opened with plot.osm.polygons)
}
\description{
plot.osm.polygons initiates a graphics device to which add.osm.polygons ()
adds successive layers of polygons generated by get.osm.polygons ().
}

