% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add-axes.R
\name{add_axes}
\alias{add_axes}
\title{add_axes}
\usage{
add_axes(map, colour = "black", axis_pos = c(0.02, 0.03), alpha = 0.5,
  fontsize = 3, fontface, fontfamily)
}
\arguments{
\item{map}{A ggplot2 object to which the axes are to be added}

\item{colour}{Colour of axis (determines colour of all elements: lines,
ticks, and labels)}

\item{axis_pos}{Positions of axes and labels relative to entire plot device}

\item{alpha}{alpha value for semi-transparent background surrounding axes and
labels (lower values increase transparency)}

\item{fontsize}{Size of axis font}

\item{fontface}{Fontface for axis labels (1:4=plain,bold,italic,bold-italic)}

\item{fontfamily}{Family of axis font (for example, 'Times')}
}
\description{
Adds axes to the internal region of an OSM plot.
}
\examples{
bbox <- get_bbox (c (-0.13, 51.5, -0.11, 51.52))
map <- plot_osm_basemap (bbox=bbox, bg="gray20")
map <- add_osm_objects (map, london$dat_BNR, col="gray40")
map <- add_axes (map)
print (map)
}

