% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add-osm-objects.R
\name{add_osm_objects}
\alias{add_osm_objects}
\title{add_osm_objects}
\usage{
add_osm_objects(obj = obj, col = "gray40", border = NA, ...)
}
\arguments{
\item{obj}{an sp SPDF or SLDF (list of polygons or lines) returned by
extract_osm_objects ()}

\item{col}{colour of polygons or lines (default = 'gray40')}

\item{border}{border colour of polygons}

\item{...}{other parameters to be passed to polygons (such as border),
lines (such as lwd, lty), or points (such as pch, cex)}
}
\value{
nothing (adds to graphics.device opened with plot_osm_basemap())
}
\description{
Adds layers of spatial objects (polygons, lines, or points generated by
extract_osm_objects ()) to a graphics object initialised with
plot_osm_basemap().
}
\examples{
plot_osm_basemap (bbox=get_bbox (c (-0.15, 51.5, -0.1, 51.52)), col="gray20")
add_osm_objects (london$dat_BNR, col="gray40") # non-residential buildings
}

