% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/click-map.R
\name{click_map}
\alias{click_map}
\title{click_map}
\usage{
click_map()
}
\value{
A data frame containing coordinates of convex hull boundary.
}
\description{
Translates clicks on a map into a convex hull object which can be passed to
add_osm_groups (). Clicking is stopped when same point on map is
clicked twice.
}

