% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colour-mat.R
\name{colour_mat}
\alias{colour_mat}
\title{colour_mat}
\usage{
colour_mat(n = c(10, 10), cols = NULL, rotate = NULL, plot = FALSE)
}
\arguments{
\item{n}{number of rows and columns (default = 10; if length 2, then
dimensions of rectangle).}

\item{cols}{vector of length >= 4 of colors (example, default = rainbow
(4), or from RColorBrewer, brewer.pal (4, 'Set1')). cols are wrapped
clockwise around the corners from top left to bottom left.}

\item{rotate}{rotates the entire colour matrix by the specified angle (in
degrees).}

\item{plot}{plots the colour matrix (default FALSE)}
}
\value{
matrix of colours
}
\description{
Generates a 2D matrix of graduated colours.
}
\examples{
library (osmplotr)
cm <- colour_mat (n=20, cols=rainbow(4), rotate=90, plot=TRUE)
}

