% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect-highways.R
\name{connect_highways}
\alias{connect_highways}
\title{connect_highways}
\usage{
connect_highways(highways, bbox, plot = FALSE)
}
\arguments{
\item{highways}{A vector of highway names passed directly to the Overpass
API. Wildcards and whitespaces are '.'; for other options see online help for
the overpass API.}

\item{bbox}{the bounding box for the map.  A 2-by-2 matrix of 4 elements with
columns of min and max values, and rows of x and y values.}

\item{plot}{If TRUE, then all OSM data for each highway is plotted and the
final cycle overlaid.}
}
\value{
A single data.frame containing the lat-lon coordinates of the cyclic
line connecting all given streets.
}
\description{
Takes a list of highways names which must enclose an internal area, and returns
a SpatialLines object containing a sequence of OSM nodes which cyclically
connect all highways. Will fail if the streets do not form a cycle.
}
\note{
This function can not be guaranteed failsafe owing both to the
inherently unpredictable nature of OpenStreetMap, as well as to the unknown
relationships between named highways. The `plot` option enables problematic
cases to be examined and hopefully resolved.  The function is still
experimental, so please help further improvements by reporting any problems!
}
\examples{
\dontrun{
bbox <- get_bbox (c(-0.15,51.5,-0.10,51.52)) # Central London, U.K.
highways <- c ('Monmouth.St', 'Short.?s.Gardens', 'Endell.St', 'Long.Acre',
               'Upper.Saint.Martin')
# Note that dots signify "anything", including whitespace and apostrophes, and
# that '?' denotes optional previous character and so here matches both
# "Shorts Gardens" and "Short's Gardens"
highways1 <- connect_highways (highways=highways, bbox=bbox, plot=TRUE)
highways <- c ('Endell.St', 'High.Holborn', 'Drury.Lane', 'Long.Acre')
highways2 <- connect_highways (highways=highways, bbox=bbox, plot=TRUE)
}
}

