% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanplotr.R
\docType{data}
\name{datA}
\alias{datA}
\title{datA OSM polygons for amenities in central London}
\format{A SpatialPolygonsDataFrame with 1,157 polygons}
\usage{
data(datA)
}
\description{
A SpatialPolygonsDataFrame containing OpenStreetMap polygons for amenities
(key=amenity) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).
}
\keyword{datasets}

