% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanplotr.R
\docType{data}
\name{datBU}
\alias{datBU}
\title{datBU}
\format{A SpatialPolygonsDataFrame with 6,209 polygons}
\usage{
data(datBU)
}
\description{
A SpatialPolygonsDataFrame containing OpenStreetMap polygons for buildings
(key=building) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).
}
\keyword{datasets}

