% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanplotr.R
\docType{data}
\name{datG}
\alias{datG}
\title{A SpatialPolygonsDataFrame containing OpenStreetMap polygons for grassed
areas (key=landuse, value=grass) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).}
\format{A SpatialPolygonsDataFrame with 50 polygons}
\usage{
data(datG)
}
\description{
A SpatialPolygonsDataFrame containing OpenStreetMap polygons for grassed
areas (key=landuse, value=grass) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).
}
\keyword{datasets}

