% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanplotr.R
\docType{data}
\name{datH}
\alias{datH}
\title{A SpatialLinesDataFrame containing OpenStreetMap lines for all highways
(key=highway) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).}
\format{A SpatialLinesDataFrame with 4,517 lines}
\usage{
data(datH)
}
\description{
A SpatialLinesDataFrame containing OpenStreetMap lines for all highways
(key=highway) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).
}
\keyword{datasets}

