% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanplotr.R
\docType{data}
\name{datN}
\alias{datN}
\title{datN}
\format{A SpatialPolygonsDataFrame with 36 polygons}
\usage{
data(datN)
}
\description{
A SpatialPolygonsDataFrame containing OpenStreetMap polygons for natural
areas (key=natural) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).
}
\keyword{datasets}

