% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urbanplotr.R
\docType{data}
\name{datP}
\alias{datP}
\title{A SpatialPolygonsDataFrame containing OpenStreetMap polygons for parks
(key=leisure, value=park) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).}
\format{A SpatialPolygonsDataFrame with 49 polygons}
\usage{
data(datP)
}
\description{
A SpatialPolygonsDataFrame containing OpenStreetMap polygons for parks
(key=leisure, value=park) in a small part of central London, U.K.
(bbox = -0.15, 51.5, -0.1, 51.52).
}
\keyword{datasets}

