% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.highway.R
\name{extract.highway}
\alias{extract.highway}
\title{extract.highway}
\usage{
extract.highway(name = "", bbox = c(-0.15, 51.5, -0.1, 51.52))
}
\arguments{
\item{bbox}{= the bounding box within which all key-value objects should be
downloaded. Default is a small part of central London.}

\item{name:}{Name of highway. Lines components are return for *any* OSM way
with a partially-matched. Both wildcards and whitespace should be represented
by ".".}
}
\value{
SpatialLinesDataFrame containing the highway
}
\description{
Extracts an OpenStreetMap highway by name, within the give bounding box.
}

