% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.osm.polygons.R
\name{extract.osm.polygons}
\alias{extract.osm.polygons}
\title{extract.osm.polygons}
\usage{
extract.osm.polygons(dat = NULL, key = "building", value = NULL,
  bbox = NULL)
}
\arguments{
\item{dat}{= raw XML data returned from get.osm.data () for the given object
type (if NULL, raw data are downloaded here, for which bbox must be provided).}

\item{type:}{currently recognises "building", "waterway", "natural", "grass",
"park", "amenity", "shop", "boundary", and "highway". Others will be simply
passed to "way[type=*]", which may or may not work.}
}
\value{
sp object containing all polygons within bbox of given type.
}
\description{
For the given vector of structure types returns a list of unambiguous
prefixes to be appended to the objects returned by get.osm.data ()
}

