% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract-highways.R
\name{extract_highways}
\alias{extract_highways}
\title{extract_highways}
\usage{
extract_highways(highway_names = NULL, bbox = NULL)
}
\arguments{
\item{highway_names}{= A vector of highway names passed directly to the
Overpass API. Wildcards and whitespaces are '.'; for other options see
overpass help.}

\item{bbox}{= the bounding box within which to look for highways.
Must be a vector of 4 elements (xmin, ymin, xmax, ymax).
Default is a small part of central London.}
}
\value{
A single data.frame containing the lat-lon coordinates of the cyclic
line connecting all given streets.
}
\description{
Extracts a list of named OpenStreetMap highways. OSM data are neither
structured nor ordered; this routine reduces data for each given highway to a
minimal number of discrete and sequentially ordered segments.
}

