% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.colours.R
\name{get.colours}
\alias{get.colours}
\title{get.colours}
\usage{
get.colours(structs = get.suffixes())
}
\value{
list of colours
}
\description{
Returns a list of sample colours for all object types listed in get.suffixes
which may be used to plot an OSM map. These are the default values used in
"make.osm.map".
}

