% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.osm.data.R
\name{get.osm.data}
\alias{get.osm.data}
\title{get.osm.data}
\usage{
get.osm.data(key = "building", value = NULL, bbox = c(-0.15, 51.5, -0.1,
  51.52))
}
\arguments{
\item{bbox}{= the bounding box within which all objects of the given type
should be downloaded. Default is a tiny bit of central London
(-0.15,51.5,-0.1,51.52).}

\item{key:}{OSM key to search for. (Useful keys include "building",
"waterway", "natural", "grass", "park", "amenity", "shop", "boundary", and
"highway". Others will be simply passed to "way[type=*]", which may or may
not work.)}
}
\value{
The contents of the overpass api query as an XML string
}
\description{
Downloads OSM data for selected object type from the MapQuest api.
}

