% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.xylims.R
\name{get.xylims}
\alias{get.xylims}
\title{get.xylims}
\usage{
get.xylims(obj)
}
\arguments{
\item{obj}{a spatial polygon or line data.frame returned by get.osm.objects ()}
}
\value{
list of lat-lon ranges
}
\description{
Extracts lat-lon limits from data frames of spatial polygons or lines
returned by get.osm.polygons()
}

