% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-highway-cycle.R
\name{get_highway_cycle}
\alias{get_highway_cycle}
\title{get_highway_cycle}
\usage{
get_highway_cycle(highways)
}
\arguments{
\item{highways}{A list of highways as returned by extract_highways(), each
element of which is a list of distinct segments for a particular OSM highway.}
}
\value{
A modified version of highways, extended by the addition of
connecting nodes.
}
\description{
Takes a list of OpenStreetMap highways returned by extract_highways()
and sequentially connects closest nodes of adjacent highways until the set of
highways connects to form a cycle.
}

