% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-suffixes.R
\name{get_suffixes}
\alias{get_suffixes}
\title{get_suffixes}
\usage{
get_suffixes(dat_types = c("building", "amenity", "waterway", "grass",
  "natural", "park", "highway", "boundary"))
}
\arguments{
\item{dat_types}{the vector of types of structures (defaults listed in
?get_osm_data).}
}
\value{
List of structures and corresponding prefixes.
}
\description{
For the given vector of structure types returns a list of unambiguous
suffices to be appended to the objects returned by get_osm_data (). This
function is useful only for automating map production, as illustrated in
make.osm.map().
}

