% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-xylims.R
\name{get_xylims}
\alias{get_xylims}
\title{get_xylims}
\usage{
get_xylims(obj)
}
\arguments{
\item{obj}{A spatial polygon or line data.frame returned by
get_osm_objects().}
}
\value{
list of lat-lon ranges
}
\description{
Extracts lat-lon limits from data frames of spatial polygons or lines
returned by get_osm_polygons(), or from bounding boxes of form
(xmin, ymin, xmax, ymax).
}

