% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group-osm-objects.R
\name{group_osm_objects}
\alias{group_osm_objects}
\title{group_osm_objects}
\usage{
group_osm_objects(obj = obj, groups = NULL, boundary = FALSE,
  cols = NULL, col_extra = NULL, colmat = TRUE)
}
\arguments{
\item{obj}{= an sp SPDF or SLDF (list of polygons or lines) returned by
get.osm.polygons}

\item{groups}{= a list of spatial points objects, each of which contains the
coordinates of points defining one group}

\item{boundary}{Either a single boolean value or a vector of same length as
groups specifying whether groups already define a boundary (TRUE), or whether
a convex hull boundary should be constructed from groups (FALSE).}

\item{cols}{= 1. Either a vector of >= 4 colours passed to colour_mat (is
colmat=T) to arrange as a 2-D map of visually distinct colours (NULL default
uses rainbow colours), or 2. If !colmat, a vector of the same length as
groups specifying individual colours for each.}

\item{col_extra}{if NULL, then any polygons *NOT* within the convex hulls are
assigned to nearest group and coloured accordingly; if NOT NULL, then any
polygons not within groups are coloured this colour.}

\item{colmat}{= TRUE generates colours according to get.colours, otherwise
the colours of groups are specified directly by the vector of cols.}
}
\value{
nothing (adds to graphics.device opened with plot.osm.basemap)
}
\description{
Plots spatially distinct groups of osm objects in different colours. OSM
objects are attributed to groups based on mean coordinates, so this routine
work best for SpatialPolygons, while may give odd results for SpatialLines.
}

