% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-highways.R
\name{haversand}
\alias{haversand}
\title{haversand}
\usage{
haversand(way1, way2)
}
\arguments{
\item{way1}{A matrix or data frame of spatial coordinates}

\item{way2}{A matrix or data frame of spatial coordinates}
}
\value{
Vector of 3 elements: numbers of elements in (way1, way2)
corresponding to minimal distance, and the distance itself.
}
\description{
Returns the minimal Haversand distance between 2 ways, along with the
element numbers in each way corresponding to that minimal distance
}

