% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-osm-objects.R
\name{list2df}
\alias{list2df}
\title{list2df}
\usage{
list2df(xy, islines = FALSE)
}
\arguments{
\item{xy}{A list of coordinates extracted from an sp object}

\item{islines}{Set to TRUE for spatial lines, otherwise FALSE}
}
\value{
data frame
}
\description{
Converts lists of coordinates to single data frames
}

