% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order.lines.R
\name{order.lines}
\alias{order.lines}
\title{order.lines}
\usage{
order.lines(spLines)
}
\arguments{
\item{spLines}{A SpatialLinesDataFrame returned from extract.osm.objects}
}
\value{
nothing (adds to graphics.device opened with plot.osm.basemap)
}
\description{
Accepts a SpatialLinesDataFrame representing an OpenStreetMap line object
such as a highway. The list items of these objects are arbitrarily organised
within OpenStreetMap. This function orders the components, returning a single
data frame with coordinates ordered sequentially along the line.
}

