% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osm-structures.R
\name{osm_structures}
\alias{osm_structures}
\title{osm_structures}
\usage{
osm_structures(structures = c("building", "amenity", "waterway", "grass",
  "natural", "park", "highway", "boundary", "tree"), col_scheme = "dark")
}
\arguments{
\item{structures}{The vector of types of structures (defaults listed in
\code{?extract_osm_objects}).}

\item{col_scheme}{Colour scheme for the plot (current options include "dark"
and "light")}
}
\value{
List of structures, corresponding prefixes, and colours.
}
\description{
For the given vector of structure types returns a \code{data.frame}
containing two columns of corresponding OpenStreetMap key-value pairs, one
column of unambiguous suffices to be appended to the objects returned by
\code{get_osm_data}, and one column specifying colours. This data frame may be
subsequently modified as desired, and ultimately passed to
\code{make_osm_map} to automate map production.
}

