% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.osm.basemap.R
\name{plot.osm.basemap}
\alias{plot.osm.basemap}
\title{plot.osm.basemap}
\usage{
\method{plot}{osm.basemap}(xylims = xylims, filename = NULL,
  bg = "gray20", graphic.device = "png", ...)
}
\arguments{
\item{xylims}{= range to be plotted as returned from get.xylims ()}

\item{filename}{= name of plot file; default=NULL plots to screen device (low
quality and likely slow)}

\item{bg}{= background colour of map (default = "gray20")}

\item{graphic.device}{= "png" (default), "jpeg", "png", or "tiff"}

\item{...}{other parameters to be passed to graphic device (such as width and
height; see ?png, for example, for details)}
}
\value{
nothing (generates file of specified type).
}
\description{
Generates a base OSM plot ready for polygon and line objects to be overlain
with add.osm.polygons () and add.osm.lines ()
}

