% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-osm-basemap.R
\name{plot_osm_basemap}
\alias{plot_osm_basemap}
\title{plot_osm_basemap}
\usage{
plot_osm_basemap(xylims = xylims, filename = NULL, width = 640,
  bg = "gray20", graphic.device = "png", ...)
}
\arguments{
\item{xylims}{= Latitude-longitude range to be plotted as returned from
get_xylims ()}

\item{filename}{= name of plot file; default=NULL plots to
screen device (low
quality and likely slow)}

\item{width}{= width of graphics file (in px; default 480).}

\item{bg}{= background colour of map (default = "gray20")}

\item{graphic.device}{= "png" (default), "jpeg", "png", or "tiff"}

\item{...}{other parameters to be passed to graphic device (such as width and
height; see ?png, for example, for details)}
}
\value{
nothing (generates file of specified type).
}
\description{
Generates a base OSM plot ready for polygon and line objects to be overlain
with add_osm_polygons () and add_osm_lines (). NOTE: Graphics files must be
closed after finishing map with dev.off() or graphics.off(). Files size may
be adjusted by width only: height is automatically calculated from the aspect
ratio of the bounding box.
}

