% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_land.R
\name{sf_land}
\alias{sf_land}
\title{Create land polygons for plotting}
\usage{
sf_land(osmd, bbox)
}
\arguments{
\item{osmd}{result of a coastline osm query}

\item{bbox}{the box associated with the query}
}
\value{
a simple feature collection
}
\description{
Create land polygons for plotting
}
