% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-highways.R
\name{shortest_way}
\alias{shortest_way}
\title{shortest_way}
\usage{
shortest_way(way, node_from, node_to)
}
\arguments{
\item{way}{A single highway (as a list of OSM components)}

\item{node_from}{The ID of a node in way from which to calculate the
shortest_path}

\item{node_to}{The ID of a node in way towards which to calculate the
shortest_path}
}
\value{
Shortest path as list of nodal IDs (or NULL if no shortest path)
}
\description{
returns the shortest path between node_from and node_to (both as names
of nodes), or NULL if node_from and node_to are not connected.
}

