% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/streets2polygon.R
\name{streets2polygon}
\alias{streets2polygon}
\title{streets2polygon}
\usage{
streets2polygon(highways = NULL, bbox = NULL, force_join = TRUE,
  plot = FALSE, lwds = c(1, 3), cols = c("black", "red"))
}
\arguments{
\item{highways}{= A vector of highway names passed directly to the Overpass
API. Wildcards and whitespaces are '.'; for other options see overpass help.}

\item{bbox}{= the bounding box within which to look for highways.
Must be a vector of 4 elements (xmin, ymin, xmax, ymax).
Default is a small part of central London.}

\item{force_join}{forces streets to form a cycle by sucessively joining
closest segments of adjacent streets}

\item{plot}{if TRUE, then all OSM data for each highway is plotted (with
lwds[1], cols[1]), with the final cycle overlaid (with lwds[2], cols[2]).}

\item{lwds}{= line widths for (all highways, final cycle)}

\item{cols}{= line colours for (all highways, final cycle)}
}
\value{
A single data.frame containing the lat-lon coordinates of the cyclic
line connecting all given streets.
}
\description{
takes a list of street names which must enclose an internal area, and returns
a SpatialLines object containing a sequence of OSM nodes which cyclically
connect all streets. Will fail if the streets do not form a cycle.
}

