% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add-axes.R
\name{add_axes}
\alias{add_axes}
\title{add_axes}
\usage{
add_axes(axis_col = "white", axis_ps = 8, axis_pos = c(0.01, 0.03))
}
\arguments{
\item{axis_col}{Colour of axis (determines colour of all elements: lines,
ticks, and labels)}

\item{axis_ps}{Point size of axis font (default 8)}

\item{axis_pos}{Relative positions of axes and labels (default=c(0.01,0.03))}
}
\description{
Adds axes to the internal region of an OSM plot.
}
\examples{
plot_osm_basemap (bbox=get_bbox (c (-0.15, 51.5, -0.1, 51.52)), col="gray20")
add_osm_objects (london$dat_BNR, col="gray40") # non-residential buildings
add_axes ()
}

