% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add-osm-surface.R
\name{add_osm_surface}
\alias{add_osm_surface}
\title{add_osm_suface}
\usage{
add_osm_surface(obj = obj, dat = NULL, method = "idw", bg = NULL,
  cols = terrain.colors(30), border = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{obj}{an sp SPDF or SLDF (list of polygons or lines) returned by
extract_osm_objects ()}

\item{dat}{A matrix or data frame of 3 columns (x, y, z), where (x, y) are
(longitude, latitude), and z are the values to be interpolated}

\item{method}{Either 'idw' (Inverse Distance Weighting as spatstat::idw;
default), otherwise uses 'Gaussian' for kernel smoothing (as
spatstat::Smooth.ppp)}

\item{bg}{If specified, OSM objects outside the convex hull surrounding 'dat'
are plotted in this colour, otherwise they are included in the interpolation
(which will generally be inaccurate for peripheral values)}

\item{cols}{Vector of colours for shading z-values (for example,
'terrain.colors (30)')}

\item{border}{Plot border? (For SpatialPolygons only)}

\item{verbose}{If TRUE, provides notification of progress}

\item{...}{other parameters to be passed to polygons, lines (such as lwd,
lty), or points (such as pch, cex)}
}
\value{
nothing (adds to graphics.device opened with plot_osm_basemap())
}
\description{
Adds a colour-coded surface of spatial objects (polygons, lines, or points
generated by extract_osm_objects ()) to a graphics object initialised with
plot_osm_basemap(). The surface is spatially interpolated between the values
given in 'dat', which has to be a matrix or data.frame of 3 colums (x, y, z),
where (x,y) are (longitude, latitude), and z are the values to be
interpolated. Interpolation uses spatstat::Smoothing.ppp, which applies a
Gaussian kernel smoother optimised to the given data, and is effectively
non-parametric.
}
\note{
Spatial smoothing is *interpolative* and so values beyond the bounding
polygon of 'dat' will generally be unreliable. Points beyond the bounding
polygon are only included in the interpolation if 'bg' is NA or NULL.
}

