% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect-highways.R
\name{connect_highways}
\alias{connect_highways}
\title{connect_highways}
\usage{
connect_highways(highways = NULL)
}
\arguments{
\item{highways}{= A vector of highways returned by extract_highways}
}
\value{
A modified version of highways, extended by the addition of
connecting nodes.
}
\description{
Takes a list of OpenStreetMap highways returned by extract_highways and
sequentially connects closest nodes of adjacent highways until the set of
highways connects to form a cycle.
}

