% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract-highway.R
\name{extract_highway}
\alias{extract_highway}
\title{extract_highway}
\usage{
extract_highway(name = "", bbox = NULL)
}
\arguments{
\item{name}{Name of highway. Lines components are return for *any* OSM way
with a partially-matched. Both wildcards and whitespace should be represented
by '.'.}

\item{bbox}{the bounding box within which to look for highways.  Must be a
vector of 4 elements (xmin, ymin, xmax, ymax).}
}
\value{
SpatialLinesDataFrame containing the highway

SpatialLinesDataFrame containing the highway
}
\description{
Extracts an OpenStreetMap highway by name, within the given bounding box.
}

