% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-highway.R
\name{extract_highway}
\alias{extract_highway}
\title{extract_highway}
\usage{
extract_highway(name = "", bbox)
}
\arguments{
\item{name}{Name of highway. Lines components are return for *any* OSM way
with a partially-matched. Both wildcards and whitespace should be represented
by '.'.}

\item{bbox}{the bounding box for the map.  A 2-by-2 matrix of 4 elements with
columns of min and max values, and rows of x and y values.}
}
\value{
SpatialLinesDataFrame containing the highway

SpatialLinesDataFrame containing the highway
}
\description{
Extracts an OpenStreetMap highway by name, within the given bounding box.
}

