% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract-highways.R
\name{extract_highways}
\alias{extract_highways}
\title{extract_highways}
\usage{
extract_highways(highway_names = NULL, bbox = NULL)
}
\arguments{
\item{highway_names}{A vector of highway names passed directly to the
Overpass API. Wildcards and whitespaces are '.'; for other options see
overpass help.}

\item{bbox}{the bounding box for the map.  A 2-by-2 matrix of 4 elements with
columns of min and max values, and rows of x and y values.}
}
\value{
A list of highways matching 'highway_names', each element of which is
a list of distinct components for the given highway.

A list of 2 components:
\enumerate{
 \item obj: A data frame of sp objects
 \item warn: Any warnings produced in downloading the data
}
}
\description{
Extracts a list of named OpenStreetMap highways. OSM data are neither
structured nor ordered; this routine reduces data for each given highway to a
minimal number of discrete and sequentially ordered segments. These segments
may or may not connect, yet may be connected at their nearest points with
connect_highways().
}

