% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract-osm-objects.R
\name{extract_osm_objects}
\alias{extract_osm_objects}
\title{extract_osm_objects}
\usage{
extract_osm_objects(key = "building", value = NULL, bbox = NULL,
  extra_pairs = NULL)
}
\arguments{
\item{key}{OSM key to search for. Useful keys include 'building',
'waterway', 'natural', 'grass', 'park', 'amenity', 'shop', 'boundary', and
'highway'. Others will be passed directly to the overpass API and may not
necessarily return results.}

\item{value}{OSM value to match to key. If NULL, all keys will be returned.
Negation is specified by '!value'.}

\item{bbox}{the bounding box within which all key-value objects should be
downloaded.  Must be a vector of 4 elements (xmin, ymin, xmax, ymax).}

\item{extra_pairs}{A list of additional key-value pairs to be passed
to the overpass API.}
}
\value{
Data frame of either spatial polygons or spatial lines
}
\description{
Downloads OSM XML objects and extracts 'sp' polygons or lines.  Requires
conversion to osmar object which can be quite slow, as can final conversion
to sp object for large numbers of objects
}

