% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.xy.lims.R
\name{get.xylims}
\alias{get.xylims}
\title{get.xylims}
\usage{
get.xylims(poly, aspect1 = TRUE)
}
\arguments{
\item{poly}{a polygon object returned by get.osm.polygons ()}

\item{aspect1}{}
}
\value{
list of lat-lon ranges
}
\description{
Extracts lat-lon limits from a set of polygons returned by get.osm.polygons()
}

