% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bbox.R
\name{get_bbox}
\alias{get_bbox}
\title{get_bbox}
\usage{
get_bbox(latlon)
}
\arguments{
\item{latlon}{A vector of (longitude, latitude, longitude, latitude) values.}
}
\value{
A 2-by-2 matrix of 4 elements with columns of min and max values, and
rows of x and y values.
}
\description{
Converts a string of latitudes and longitudes into a square matrix to be
passed as a \code{bbox} argument (to \code{\link{extract_osm_objects}},
\code{\link{osm_basemap}}, or \code{\link{make_osm_map}}).
}
\examples{
bbox <- get_bbox (c (-0.15, 51.5, -0.1, 51.52))
}
