% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group.osm.objects.R
\name{group.osm.objects}
\alias{group.osm.objects}
\title{group.osm.objects}
\usage{
group.osm.objects(obj = obj, groups = NULL, cols = NULL,
  col.extra = NULL, colmat = TRUE)
}
\arguments{
\item{obj}{= an sp SPDF or SLDF (list of polygons or lines) returned by
get.osm.polygons}

\item{groups}{= a list of spatial points objects, each of which contains the
coordinates of points defining one group}

\item{cols}{= 1. Either a vector of >= 4 colours passed to colour.mat (is
colmat=T) to arrange as a 2-D map of visually distinct colours (NULL default
uses rainbow colours), or 2. If !colmat, a vector of the same length as
groups specifying individual colours for each.}

\item{col.extra}{if NULL, then any polygons *NOT* within the convex hulls are
assigned to nearest group and coloured accordingly; if NOT NULL, then any
polygons not within groups are coloured this colour.}
}
\value{
nothing (adds to graphics.device opened with plot.osm.basemap)
}
\description{
Plots spatially distinct groups of osm objects is different colours. OSM
objects are attributed to groups based on mean coordinates, so this routine
work best for SpatialPolygons, while may give odd results for SpatialLines.
}

