% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmplotr.R
\docType{data}
\name{london}
\alias{london}
\title{london}
\format{A list of spatial objects}
\description{
A list of SpatialPolygonsDataFrames (SPDF) and SpatialLinesDataFrames (SLDF)
containing OpenStreetMap polygons and lines for various OpenStreetMap
structures in a small part of central London, U.K.  (bbox = -0.15, 51.5,
-0.1, 51.52). The list includes:
\enumerate{
 \item dat_H an SLDF of non-primary highways with 3,868 lines
 \item dat_HP an SLDF of primary highways with 659 lines
 \item dat_B an SPDF of non-commerical buildings with 6,177 polygons
 \item dat_BC an SPDF of commerical buildings with 30 polygons
 \item dat_A an SPDF of amenities with 1,157 polygons
 \item dat_G an SPDF of grassed areas with 50 polygons
 \item dat_P an SPDF of parks with 49 polygons
 \item dat_N an SPDF of natural areas with 36 polygons
 \item dat_RFH an SPDF containing 1 polygon representing Royal Festival Hall
 \item dat_ST an SPDF containing 1 polygon representing 150 Stamford Street
 \item highways1 A SpatialPoints object containing 178 points representing
 the circular perimeter of ('Kingsway', 'Holborn', 'Farringdon.St', 'Strand',
 'Fleet.St', 'Aldwych')
 \item highways2 A SpatialPoints object containing 126 points representing
 the circular perimeter of ('Queen.s.Walk', 'Blackfriars', 'Waterloo',
 'The.Cut')
 \item highways2 A SpatialPoints object containing 126 points representing
 the circular perimeter of ('Regent.St', 'Oxford.St', 'Shaftesbury')
}
}
\examples{
# Data were downloaded using the following code:
bbox <- c(-0.13,51.5,-0.12,51.51) # smaller bbox for the example
#bbox <- c(-0.15,51.5,-0.1,51.52) # but this one actually used
structures <- c ('highway', 'highway', 'building', 'building', 'building',
                 'amenity', 'grass', 'park', 'natural', 'tree')
structs <- osm_structures (structures=structures, col_scheme='dark')
structs$value [1] <- '!primary'
structs$value [2] <- 'primary'
structs$suffix [2] <- 'HP'
structs$value [3] <- '!residential'
structs$value [4] <- 'residential'
structs$value [5] <- 'commercial'
structs$suffix [3] <- 'BNR'
structs$suffix [4] <- 'BR'
structs$suffix [5] <- 'BC'
london <- list ()
for (i in 1:(nrow (structs) - 1))
{
    dat <- extract_osm_objects (key=structs$key [i], value=structs$value [i],
                               bbox=bbox)
    fname <- paste0 ('dat_', structs$suffix [i])
    assign (fname, dat)
    london [[i]] <- get (fname)
    names (london)[i] <- fname
    rm (list=c(fname))
}
bbox <- c(-0.15,51.5,-0.1,51.52)
extra_pairs <- c ('name', 'Royal.Festival.Hall')
london$dat_RFH <- extract_osm_objects (key='building', extra_pairs=extra_pairs,
                                bbox=bbox)
extra_pairs <- list (c ('addr:street', 'Stamford.St'),
                     c ('addr:housenumber', '150'))
london$dat_ST <- extract_osm_objects (key='building', extra_pairs=extra_pairs,
                                bbox=bbox)
highways <- c ('Kingsway', 'Holborn', 'Farringdon.St', 'Strand',
               'Fleet.St', 'Aldwych')
#london$highways1 <- highways2polygon (highways=highways, bbox=bbox)
highways <- c ('Queen.s.Walk', 'Blackfriars', 'Waterloo', 'The.Cut')
london$highways2 <- highways2polygon (highways=highways, bbox=bbox)
highways <- c ('Regent.St', 'Oxford.St', 'Shaftesbury')
london$highways3 <- highways2polygon (highways=highways, bbox=bbox)
}
\keyword{datasets}

