% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmplotr.R
\docType{data}
\name{london}
\alias{london}
\title{london}
\format{A list of spatial objects}
\description{
A list of \code{Simple Features} (\code{sf}) \code{data.frame} objects
containing OpenStreetMap polygons, lines, and points for various
OpenStreetMap structures in a small part of central London, U.K.  (\code{bbox
= -0.13, 51.51, -0.11, 51.52}). The list includes:
\enumerate{
 \item \code{dat_H}: 974 non-primary highways as linestrings
 \item \code{dat_HP}: 159 primary highways as linestrings
 \item \code{dat_BNR}: 1,716 non-residential buildings as polygons
 \item \code{dat_BR}: 43 residential buildings as polygons
 \item \code{dat_BC}: 67 commerical buildings as polygons
 \item \code{dat_A}: 372 amenities as polygons
 \item \code{dat_P}: 13 parks as polygons
 \item \code{dat_T}: 688 trees as points
 \item \code{dat_RFH}: 1 polygon representing Royal Festival Hall 
 \item \code{dat_ST}: 1 polygon representing 150 Stamford Street
}
}
\details{
The vignette \code{basic-maps} details how these data were downloaded. Note
that these internal versions have had all descriptive data removed other than
their names, geometries, and their OSM identification numbers.
}
\keyword{datasets}
