% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make-osm-map.R
\name{make_osm_map}
\alias{make_osm_map}
\title{make_osm_map}
\usage{
make_osm_map(filename = NULL, bbox = NULL, osm_data = NULL,
  structures = osm_structures(), dat_prefix = "dat_")
}
\arguments{
\item{filename}{Name of plot file; default=NULL plots to screen device (low
quality and likely slow)}

\item{bbox}{The bounding box for the map.  Must be a vector of 4 elements
(xmin, ymin, xmax, ymax). If NULL, bbox is taken from the largest extent of
OSM objects in osm_data.}

\item{osm_data}{A list of OSM objects as returned from
\code{extract_osm_objects}.  These objects may be included in the plot
without downloading. These should all be named with the stated
\code{dat_prefix} and have suffixes as given in \code{structures}.}

\item{structures}{A data.frame specifying types of OSM structures as returned
from osm_structures, and potentially modified to alter lists of structures to
be plotted, and their associated colours. The order of structs determines the
plot order of objects.}

\item{dat_prefix}{Prefix for data structures (default 'dat_'). Final data
structures are created by appending the suffixes from \code{osm_structures}.}
}
\value{
list of OSM structures each as Spatial(Polygon/List)DataFrame, and
appended to `osm_data` (which is NULL by default).
}
\description{
Makes an entire OSM map for the given bbox by downloading all data.  This
function can take quite some time (tens of minutes) to execute the two
primary steps of downloading and processing the raw OSM data.  For this
reason, this function is not intended to be actually run; rather it serves to
demonstrate how all functions may be combined to generate a map.
}
\details{
Progress is dumped to screen, including time taken.
}

