% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmplotr.R
\docType{package}
\name{osmplotr}
\alias{osmplotr}
\alias{osmplotr-package}
\title{osmplotr.}
\description{
Produces customisable images of OpenStreetMap data.  Extracts OpenStreetMap
data for specified key-value pairs (e.g.  key='building') using the overpass
API. Different OSM objects can be plotted in different colours using the
function add_osm_objects().  The function group_osm_objects()
enables customised highlighting of selected regions using different graphical
schemes designed to contrast with surrounding backgrounds.
}

